+++ package.json
@@ -12,7 +12,7 @@
     "cypress:open": "cypress open",
     "cypress:run": "cypress run --headless --browser chrome",
     "cypress:ci": "yarn build && start-server-and-test start http://localhost:3000 cypress:run",
-    "typechain": "typechain --target ethers-v5 \"pages/future/dev/**/*.json\""
+    "typechain": "typechain --target ethers-v5 \"contracts/**/*.json\" --out-dir lib/generated/ethers-contract"
   },
   "dependencies": {
     "@auth0/nextjs-auth0": "^1.9.1",
