+++ pages/api/minter/verify.ts
